+++ plugins/producer_plugin/producer_plugin.cpp
@@ -647,7 +647,7 @@ class producer_plugin_impl : public std::enable_shared_from_this<producer_plugin
                                               || ( persist_until_expired && _disable_subjective_api_billing )
                                               || ( !persist_until_expired && _disable_subjective_p2p_billing );
 
-            uint32_t sub_bill = 0;
+            int64_t sub_bill = 0;
             if( !disable_subjective_billing )
                sub_bill = _subjective_billing.get_subjective_bill( first_auth, fc::time_point::now() );
 
@@ -859,7 +859,7 @@ bool producer_plugin::is_producer_key(const chain::public_key_type& key) const
   return false;
 }
 
-uint32_t producer_plugin::get_subjective_bill( const account_name& first_auth, const fc::time_point& now ) const
+int64_t producer_plugin::get_subjective_bill( const account_name& first_auth, const fc::time_point& now ) const
 {
    return my->_subjective_billing.get_subjective_bill( first_auth, now );
 }
@@ -1939,7 +1939,7 @@ bool producer_plugin_impl::process_unapplied_trxs( const fc::time_point& deadlin
                trx_deadline = deadline;
             }
             // no subjective billing since we are producing or processing persisted trxs
-            const uint32_t sub_bill = 0;
+            const int64_t sub_bill = 0;
 
             auto trace = chain.push_transaction( trx, trx_deadline, prev_billed_cpu_time_us, false, sub_bill );
             fc_dlog( _trx_failed_trace_log, "Subjective unapplied bill for ${a}: ${b} prev ${t}us", ("a",first_auth)("b",prev_billed_cpu_time_us)("t",trace->elapsed));
